+++ commands/settings/generate.js
@@ -3,6 +3,7 @@ module.exports = {
     name: "generate",
     description: "generate a new settings\'s channel (if there isn\'t another one)",
     needPerm: true,
+    permissions: ["MANAGE_CHANNELS"],
     run: async (e, vars, args) => {
         let settings  = await getSettings(e.guild.id, vars, true)
         if(settings.channelID) {